refactor(character, utils): change folder name
