+++ src/app/character/service/character.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { fetchCharactersUrl } from 'src/app/utils/characterApi';
+import { fetchCharactersUrl } from 'src/app/utils/api/characterApi';
 
 @Injectable({
   providedIn: 'root',
